feat/refactor-folder-structure
