+++ src/components/styles/components.scss
@@ -1,28 +1,3 @@
-.palette {
-    &__containter {
-        border: transparent hsl(0, 0%, 0%) 10px;
-        border-radius: 20px;
-        display: flex;
-
-    }
-
-    &__colour {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        height: 9em;
-        width: 6em;
-
-        &:hover .palette__hex {
-            display: block;
-        }
-    }
-
-    &__hex {
-        display: none;
-    }
-}
-
 .header {
     display: flex;
     justify-content: space-between;
@@ -31,17 +6,17 @@
     padding-left: 2em;
     padding-right: 2em;
 
-    &__logo {
-        width: 200px;
-    }
-
-    &__title {
-        font-size: 30px;
-        margin: 0;
+    &__content {
+        padding-bottom: 1em;
 
         @media screen and (max-width: '568px') {
             text-align: center;
+            padding-bottom: 0;
         }
+    }
+
+    &__logo {
+        width: 200px;
     }
 
     &__subtitle {
@@ -53,16 +28,12 @@
         padding: 0;
         padding-top: 1em;
     }
-
-
 }
 
 .nav {
     display: flex;
     justify-content: flex-end;
     align-items: center;
-    padding: 1em;
-
 
     &__buttons {
         display: flex;
@@ -75,66 +46,4 @@
             padding: 0;
         }
     }
-}
-
-.picker {
-    display: flex;
-    justify-content: center;
-    align-items: start;
-    height: 100%;
-    width: 100%;
-
-    &__container {
-        display: flex;
-        width: 100%;
-    }
-}
-
-.colour {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    min-height: 87.8vh;
-    height: 100%;
-    width: 100%;
-
-    &__content {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        width: 100%;
-        position: relative;
-    }
-
-    &__hex-code {
-        // cursor: pointer;
-    }
-
-    &__actions {
-        height: 2em;
-    }
-
-    &__tools {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        width: 100%;
-    }
-
-    &__hover {
-        width: 10em;
-        height: 2em;
-    }
-
-    &__lock-icon {
-        margin-top: 2em;
-        height: 2em;
-    }
-}
-
-.bar-buttons {
-    display: flex;
-    // gap: 0.5em;
 }
