+++ src/pages/TrendingPage/components/PaletteDropMenu.jsx
